// crawler/src/database/hasuraAPI/graphql/generated.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  executionstatus: any;
  time: any;
};
// ...
export type Records_Insert_Input = {
  active?: InputMaybe<Scalars['Boolean']>;
  boundary?: InputMaybe<Scalars['String']>;
  executions?: InputMaybe<Executions_Arr_Rel_Insert_Input>;
  id?: InputMaybe<Scalars['Int']>;
  label?: InputMaybe<Scalars['String']>;
  nodes?: InputMaybe<Nodes_Arr_Rel_Insert_Input>;
  periodicity?: InputMaybe<Scalars['Int']>;
  tags_records_relations?: InputMaybe<Tags_Records_Relations_Arr_Rel_Insert_Input>;
  url?: InputMaybe<Scalars['String']>;
};
// ...
export type Records_Obj_Rel_Insert_Input = {
  data: Records_Insert_Input;
  /** upsert condition */
  on_conflict?: InputMaybe<Records_On_Conflict>;
};

/** on_conflict condition type for table "records" */
export type Records_On_Conflict = {
  constraint: Records_Constraint;
  update_columns?: Array<Records_Update_Column>;
  where?: InputMaybe<Records_Bool_Exp>;
};
// ...
export type Tags_Insert_Input = {
  id?: InputMaybe<Scalars['Int']>;
  tag_name?: InputMaybe<Scalars['String']>;
  tags_records_relations?: InputMaybe<Tags_Records_Relations_Arr_Rel_Insert_Input>;
};
// ...
export type Tags_Obj_Rel_Insert_Input = {
  data: Tags_Insert_Input;
  /** upsert condition */
  on_conflict?: InputMaybe<Tags_On_Conflict>;
};

/** on_conflict condition type for table "tags" */
export type Tags_On_Conflict = {
  constraint: Tags_Constraint;
  update_columns?: Array<Tags_Update_Column>;
  where?: InputMaybe<Tags_Bool_Exp>;
};
// ...
export type Tags_Records_Relations_Insert_Input = {
  id?: InputMaybe<Scalars['Int']>;
  record?: InputMaybe<Records_Obj_Rel_Insert_Input>;
  record_id?: InputMaybe<Scalars['Int']>;
  tag?: InputMaybe<Tags_Obj_Rel_Insert_Input>;
  tag_id?: InputMaybe<Scalars['Int']>;
};
// ...
export type AllRecordsQuery = { __typename?: 'query_root', records: Array<{ __typename?: 'records', id: number, active: boolean, boundary: string, label: string, periodicity: number, url: string, tags: Array<{ __typename?: 'tags_records_relations', tag: { __typename?: 'tags', tag_name: string, id: number } }> }> };

export type GetRecordQueryVariables = Exact<{
  id?: InputMaybe<Scalars['Int']>;
}>;
// ...
export type InsertRecordMutation = { __typename?: 'mutation_root', insert_records_one?: { __typename?: 'records', id: number } | null };

export type UpdateRecordMutationVariables = Exact<{
  id?: InputMaybe<Scalars['Int']>;
  dataToUpdate?: InputMaybe<Records_Set_Input>;
}>;


export type UpdateRecordMutation = { __typename?: 'mutation_root', update_records?: { __typename?: 'records_mutation_response', affected_rows: number } | null };

export type InsertTagsRecordRelationsMutationVariables = Exact<{
  objects?: InputMaybe<Array<Tags_Records_Relations_Insert_Input> | Tags_Records_Relations_Insert_Input>;
}>;


export type InsertTagsRecordRelationsMutation = { __typename?: 'mutation_root', insert_tags_records_relations?: { __typename?: 'tags_records_relations_mutation_response', affected_rows: number } | null };
// ...
export type UpdateRecordRelationsByRecordIdsMutation = { __typename?: 'mutation_root', delete_tags_records_relations?: { __typename?: 'tags_records_relations_mutation_response', affected_rows: number } | null, insert_tags_records_relations?: { __typename?: 'tags_records_relations_mutation_response', affected_rows: number } | null };